[FE] feat : Moim 가입시 페이지 레이아웃 컴포넌트화
