+++ frontend/moim_front/src/router/index.jsx
@@ -1,6 +1,6 @@
 // React
 import { useEffect } from "react";
-import { Routes, Route, useNavigate } from "react-router-dom";
+import { Routes, Route, useNavigate, useParams } from "react-router-dom";
 
 // Pages : Home
 import HomePage from "../pages/Home/Home";
@@ -32,6 +32,7 @@ import MainLayout from "../components/MainLayout";
 import ManageLayout from "../assets/ManageLayout";
 import ManageMoimPage from "../pages/ManageMoimPage";
 import ManageProfilePage from "../pages/ManageProfilePage";
+import MoimLayout from "../components/MoimLayout";
 
 /**
  * Basic Router
@@ -61,25 +62,28 @@ function Router() {
       checkSession();
     }
   }, [navigate]);
+  const { MoimId } = useParams();
 
   return (
     <Routes>
       <Route element={<MainLayout />}>
         <Route path="/" element={<HomePage />} />
-        <Route path="/join/:MoimId" element={<MoimJoinPage />} />
         <Route path="/detailed/:MoimId" element={<MoimDetailPage />} />
         <Route path="/meeting/:MoimId" element={<MeetingCreation />} />
+
+        <Route element={<MoimLayout />}>
+          <Route path="/join/:MoimId" element={<MoimJoinPage />} />
+          <Route
+            path="/meeting/:moimId/:meetingId"
+            element={<MeetingDetailPage />}
+          />
+        </Route>
       </Route>
 
       <Route element={<ManageLayout />}>
         <Route path="/manage" element={<ManageProfilePage />} />
         <Route path="/manage/moim" element={<ManageMoimPage />} />
       </Route>
-
-      <Route
-        path="/meeting/:moimId/:meetingId"
-        element={<MeetingDetailPage />}
-      />
 
       <Route path="/login" element={<LoginPage />} />
       <Route path="/register" element={<RegisterPage />} />
